feat: 학생 식당 운영 시간 멤버변수 수정 및 추가
